fix: rename test folder to testUtilities
