+++ core/src/app.test.ts
@@ -1,7 +1,7 @@
-import { MockFactory } from "./test/mockFactory";
+import { MockFactory } from "./testUtilities/mockFactory";
 import { CloudContext } from "./cloudContext";
 import { App } from "./app";
-import { TestContext } from "./test/testContext";
+import { TestContext } from "./testUtilities/testContext";
 
 const errorMiddleware = (spy: Function) => async (
   context: CloudContext
