+++ core/src/app.ts
@@ -1,7 +1,7 @@
 import { Guard } from "./common/guard";
 import { ensurePromise } from "./common/util";
 import { CloudContainer, CloudModule, ComponentType } from "./cloudContainer";
-import { TestModule } from "./test/testModule";
+import { TestModule } from "./testUtilities/testModule";
 import { CloudContext } from "./cloudContext";
 
 /**
