+++ core/src/middleware/serviceMiddleware.test.ts
@@ -1,8 +1,8 @@
 import { CloudContext } from "../cloudContext";
 import { CloudContainer } from "../cloudContainer";
-import { TestModule } from "../test/testModule";
-import { TestCloudService } from "../test/testCloudService";
-import { TestContext } from "../test/testContext";
+import { TestModule } from "../testUtilities/testModule";
+import { TestCloudService } from "../testUtilities/testCloudService";
+import { TestContext } from "../testUtilities/testContext";
 import { ServiceMiddleware } from "./serviceMiddleware";
 
 describe("ServiceMiddleware should", () => {
