+++ core/src/middleware/storageMiddleware.test.ts
@@ -1,4 +1,4 @@
-import { MockFactory } from "../test/mockFactory";
+import { MockFactory } from "../testUtilities/mockFactory";
 import { App, Handler } from "../app";
 import { CloudContext } from "../cloudContext";
 import { StorageMiddleware } from "./storageMiddleware";
